backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
openai-whisper==20231117
ffmpeg-python==0.2.0
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings>=2.1.0
aiofiles==23.2.1
torch
edge-tts>=6.1.0
redis>=5.0.0
PyJWT>=2.8.0
